Move `GraphResolveDepths` into its own module
