+++ packages/graph/hash_graph/lib/graph/src/ontology/mod.rs
@@ -158,45 +158,6 @@ where
     serde_json::Value::from(ontology_type.clone()).serialize(serializer)
 }
 
-/// Distance to explore when querying a rooted subgraph in the ontology.
-///
-/// Ontology records may have references to other records, e.g. a [`PropertyType`] may reference
-/// other [`PropertyType`]s or [`DataType`]s. The depths provided alongside a query specify how many
-/// steps to explore along a chain of references _of a certain kind of type_. Meaning, any chain of
-/// property type references will be resolved up to the depth given for property types, and *each*
-/// data type referenced in those property types will in turn start a 'new chain' whose exploration
-/// depth is limited by the depth given for data types.
-///
-/// A depth of `0` means that no references are explored for that specific kind of type.
-///
-/// [`DataType`]: type_system::DataType
-/// [`PropertyType`]: type_system::PropertyType
-///
-/// # Example
-///
-/// - `EntityType1` references \[`EntityType2`, `PropertyType1`]
-/// - `EntityType2` references \[`PropertyType2`]
-/// - `PropertyType1` references \[`DataType2`]
-/// - `PropertyType2` references \[`PropertyType3`, `DataType1`]
-/// - `PropertyType3` references \[`PropertyType4`, `DataType3`]
-/// - `PropertyType4` references \[`DataType3`]
-///
-/// If a query on `EntityType1` is made with the following depths:
-/// - `entity_type_query_depth: 1`
-/// - `property_type_query_depth: 3`
-/// - `data_type_query_depth: 1`
-///
-/// Then the returned subgraph will be:
-/// - `referenced_entity_types`: \[`EntityType2`]
-/// - `referenced_property_types`: \[`PropertyType1`, `PropertyType2`, `PropertyType3`]
-/// - `referenced_data_types`: \[`DataType1`, `DataType2`]
-///
-/// ## The idea of "chains"
-///
-/// When `EntityType2` is explored its referenced property types get explored. The chain of
-/// _property type_ references is then resolved to a depth of `property_type_query_depth`.
-pub type OntologyQueryDepth = u8;
-
 #[derive(Debug, Clone, PartialEq, Eq, Serialize, ToSchema)]
 pub struct PersistedDataType {
     #[schema(value_type = VAR_DATA_TYPE)]
